+++ .gitignore
@@ -4,3 +4,4 @@
 /NikiConnectAPI.Test/bin
 /NikiConnectAPI.Test/obj
 /packages
+/NikiConnectAPI.Test/.vs
