+++ NikiConnectAPI.Lib/Models/Auth/TokenResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace NikiConnectAPI.Lib.Models.Auth
+{
+    public class TokenResponse
+    {
+        public Token Token { get; set; }
+        public TokenError TokenError { get; set; }
+    }
+}
