+++ NikiConnectAPI.Lib/Models/Data/DataResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace NikiConnectAPI.Lib.Models.Data
+{
+    public class DataResponse<T> where T : class
+    {
+        public DataResult<T> DataResult { get; set; }
+        public DataResultError Error { get; set; }
+    }
+}
