+++ NikiConnectAPI.Test/NikiConnectAPI.Test.csproj
@@ -49,10 +49,9 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Api\Country.cs" />
+    <Compile Include="Api\ServiceModels.cs" />
     <Compile Include="Api\Auth.cs" />
     <Compile Include="App.cs" />
-    <Compile Include="UnitTest1.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
