--- NikiConnectAPI.Test/UnitTest1.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-
-namespace NikiConnectAPI.Test
-{
-    [TestClass]
-    public class UnitTest1
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-
-        }
-    }
-}
